+++ src/picbh/__init__.py
@@ -2,6 +2,6 @@
 
 __version__ = "0.0.0"
 
-from . import element, graph
+from . import cbh, element, graph
 
-__all__ = ["element", "graph"]
+__all__ = ["cbh", "element", "graph"]